Minor restructure into separate files.
