+++ lib/powersync.dart
@@ -7,6 +7,6 @@ export 'src/powersync_database.dart';
 export 'src/schema.dart';
 export 'src/connector.dart';
 export 'src/crud.dart';
-export 'src/streaming_sync.dart' show SyncStatus;
+export 'src/sync_status.dart';
 export 'src/uuid.dart';
 export 'src/open_factory.dart';
